+++ ql/time/daycounters/thirty360.cpp
@@ -37,7 +37,7 @@ namespace QuantLib {
         }
     }
 
-    BigInteger Thirty360::US_Impl::dayCount(const Date& d1,
+    int_fast32_t Thirty360::US_Impl::dayCount(const Date& d1,
                                             const Date& d2) const {
         Day dd1 = d1.dayOfMonth(), dd2 = d2.dayOfMonth();
         Integer mm1 = d1.month(), mm2 = d2.month();
@@ -49,7 +49,7 @@ namespace QuantLib {
             std::max(Integer(0),30-dd1) + std::min(Integer(30),dd2);
     }
 
-    BigInteger Thirty360::EU_Impl::dayCount(const Date& d1,
+    int_fast32_t Thirty360::EU_Impl::dayCount(const Date& d1,
                                             const Date& d2) const {
         Day dd1 = d1.dayOfMonth(), dd2 = d2.dayOfMonth();
         Month mm1 = d1.month(), mm2 = d2.month();
@@ -59,7 +59,7 @@ namespace QuantLib {
             std::max(Integer(0),30-dd1) + std::min(Integer(30),dd2);
     }
 
-    BigInteger Thirty360::IT_Impl::dayCount(const Date& d1,
+    int_fast32_t Thirty360::IT_Impl::dayCount(const Date& d1,
                                             const Date& d2) const {
         Day dd1 = d1.dayOfMonth(), dd2 = d2.dayOfMonth();
         Month mm1 = d1.month(), mm2 = d2.month();
